+++ core/src/stored/device_resource.cc
@@ -288,14 +288,10 @@ static void WarnOnZeroMaxConcurrentJobs(int max_concurrent_jobs,
                                         std::string_view name)
 {
   if (max_concurrent_jobs == 0) {
-    my_config->AddWarning(fmt::format(
-        FMT_STRING("Device {:s}: unlimited (0) 'Maximum Concurrent Jobs' (the "
-                   "default) reduces the restore peformance."),
-        name));
-    my_config->AddWarning(fmt::format(
-        FMT_STRING("Device {:s}: the default value for 'Maximum Concurrent "
-                   "Jobs' will change from 0 (unlimited) to 1 in Bareos 23."),
-        name));
+    my_config->AddWarning(
+        fmt::format(FMT_STRING("Device {:s}: unlimited (0) 'Maximum Concurrent "
+                               "Jobs' reduces the restore peformance."),
+                    name));
   }
 }