+++ core/src/stored/stored_conf.h
@@ -101,8 +101,7 @@ class StorageResource
       = nullptr; /**< Address to source connections from */
   dlist<IPADDR>* NDMPaddrs = nullptr;
   char* working_directory = nullptr; /**< Working directory for checkpoints */
-  char* pid_directory = nullptr;
-  char* plugin_directory = nullptr; /**< Plugin directory */
+  char* plugin_directory = nullptr;  /**< Plugin directory */
   alist<const char*>* plugin_names = nullptr;
   char* scripts_directory = nullptr;
   std::vector<std::string> backend_directories;